requirements.txt
anyconfig==0.12.0
arrow==1.2.3
beautifulsoup4==4.10.0
binaryornot==0.4.4
certifi==2021.10.8
chardet==5.1.0
charset-normalizer==2.0.10
checksumdir==1.2.0
click==8.0.3
click-log==0.3.2
cloudinary==1.28.1
colorama==0.4.4
commonmark==0.9.1
cookiecutter==2.1.1
copier==7.0.1
deepmerge==1.1.0
diskcache==5.4.0
docstring-parser==0.13
dunamai==1.15.0
feedgen==0.9.0
idna==3.3
importlib-metadata==4.10.1
iteration-utilities==0.11.0
Jinja2==3.1.2
jinja2-ansible-filters==1.3.2
jinja2-time==0.2.0
linkify-it-py==2.0.2
lxml==4.7.1
markata==0.7.4
Markdown==3.3.6
markdown-it-py==3.0.0
markdown2==2.4.10
MarkupSafe==2.0.1
mdit-py-plugins==0.4.0
mdurl==0.1.2
more-itertools==8.12.0
packaging==23.0
pathspec==0.9.0
Pillow==9.0.0
pluggy==1.0.0
plumbum==1.8.1
prompt-toolkit==3.0.36
pydantic==1.10.4
Pygments==2.11.2
pymdown-extensions==9.1
python-dateutil==2.8.2
python-frontmatter==1.0.0
python-slugify==7.0.0
pytz==2021.3
PyYAML==6.0
pyyaml-include==1.3
questionary==1.10.0
requests==2.27.1
rich==11.0.0
six==1.16.0
soupsieve==2.3.1
svgwrite==1.4.3
text-unidecode==1.3
textual==0.1.14
toml==0.10.2
typer==0.4.0
typing_extensions==4.4.0
uc-micro-py==1.0.2
urllib3==1.26.8
wavedrom==2.0.3.post3
wcwidth==0.2.6
zipp==3.7.0
